requirements.txt
openai
openai-whisper
moviepy==1.0.3
dotenv